yum aliases

+++ .shell_aliases_functions.sh
@@ -301,11 +301,15 @@ else
     test -d "$HOME/.local/share/Trash" && \
         alias tra='cd $HOME/.local/share/Trash'
     if [[ "$(uname)" == Linux ]]; then
-        exists apt && {
+        if exists apt;then
             alias api="sudo apt install -y"
             alias apa="sudo apt update && sudo apt dist-upgrade -y; sudo apt autoremove -y; sudo apt autoclean"
             alias apz="sudo apt update && sudo apt dist-upgrade -y; sudo apt autoremove -y; sudo apt autoclean; u8"
-        }
+        elif exists yum;then
+            alias apa="sudo yum check-update && sudo yum upgrade -y "
+            alias apz="sudo yum check-update && sudo yum upgrade -y; u8"
+        fi
+    
         alias ipt="sudo iptables --line-numbers -L"
         test -z "$distroName" && {
             distroName=$(command grep "^ID=" /etc/os-release | cut -d= -f2 | tr -d \" | head -n 1)
